name_generator/requirements.txt
asgiref==3.8.1
DAWG-Python==0.7.2
Django==4.2.17
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
